projects/BrainCancer-MRI/requirements.txt
# Core PyTorch packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing and augmentation
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0

# Machine learning and monitoring
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0
wandb>=0.15.0
mlflow>=2.0.0

# Utilities
tqdm>=4.64.0
PyYAML>=6.0
psutil>=5.9.0
GPUtil>=1.4.0

# Optional: For advanced augmentation and medical imaging
# Install via conda: conda install -c conda-forge albumentations opencv
# These packages are better installed via conda for better performance and compatibility

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
